+++ config/settings/base.py
@@ -203,6 +203,7 @@ TEMPLATES = [
                 'django.template.context_processors.tz',
                 'django.contrib.messages.context_processors.messages',
                 # Your stuff: custom template context processors go here
+                'projectroles.context_processors.urls_processor',
             ],
         },
     },